+++ Yanni/layer.h
@@ -6,7 +6,7 @@
 class Layer
 {
 public:
-	size_t Size;
+	int Size;
 	float* Inputs;
 	float** InputsBatch;
 	float* Weights;
@@ -25,10 +25,10 @@ public:
 	float* LearningRates;
 	float* Target;
 	float** TargetsBatch;
-	size_t* IndexVector;
-	size_t IndexVectorSize;
-	size_t* IndexVectorForNextLayer;
-	size_t IndexVectorForNextLayerSize;
+	int* IndexVector;
+	int IndexVectorSize;
+	int* IndexVectorForNextLayer;
+	int IndexVectorForNextLayerSize;
 	float DropOutSize;
 	bool UsingBias;
 	bool* DropoutNeurons; // if DropoutNeurons[i]==true, then Inputs[i] and Outputs[0]  do not take part in calculations
@@ -36,26 +36,26 @@ public:
 	NeuralEnums::ActivationFunction ActivationFunction;
 	NeuralEnums::LayerType LayerType;
 	Layer() {}
-	Layer(size_t size, NeuralEnums::LayerType layerType, NeuralEnums::ActivationFunction activationFunction, float bias, float dropoutSize = 0, int batchSize = 1);
+	Layer(int size, NeuralEnums::LayerType layerType, NeuralEnums::ActivationFunction activationFunction, float bias, float dropoutSize = 0, int batchSize = 1);
 
 	void CalcInputsDelegate(
 		float* prevLayerOutput,
-		size_t prevLayerSize,
+		int prevLayerSize,
 		float** prevLayerOutputBatch,
-		size_t* prevLayerIndexes,
-		size_t& prevLayerIndexVectorSize,
-		bool& training, size_t& start, size_t& end);
-	void CalcOutputsDelegate(size_t& start, size_t& end, bool& training, bool& countingRohat);
+		int* prevLayerIndexes,
+		int& prevLayerIndexVectorSize,
+		bool& training, int& start, int& end);
+	void CalcOutputsDelegate(int& start, int& end, bool& training, bool& countingRohat);
 	void CalculateInputsThreaded(
 		float* prevLayerOutput,
-		size_t prevLayerSize,
+		int prevLayerSize,
 		float** prevLayerOutputBatch,
-		size_t* prevLayerIndex,
-		size_t& prevLayerIndexVectorSize,
+		int* prevLayerIndex,
+		int& prevLayerIndexVectorSize,
 		bool& training,
-		size_t& numThreads,
+		int& numThreads,
 		std::vector<WorkerThread*>& _workers);
-	void CalculateOutputsThreaded(size_t& numThreads, bool& training, bool& countingRohat, std::vector<WorkerThread*>& _workers);
+	void CalculateOutputsThreaded(int& numThreads, bool& training, bool& countingRohat, std::vector<WorkerThread*>& _workers);
 };
 #endif
 //Layer.h
