+++ Yanni/mnistData.h
@@ -1,9 +1,9 @@
 #pragma once
 struct MnistData
 {
-	size_t setSize;
-	size_t labelSize;
-	size_t minMax[2];
+	int setSize;
+	int labelSize;
+	int minMax[2];
 	float* set;
 	float* label;
 };
