+++ Yanni/statisticFunctions.h
@@ -5,30 +5,30 @@
 #include <math.h>
 
 
-float Mean(float* x, size_t size);
+float Mean(float* x, int size);
 
-float Max(float* x, size_t xSize);
+float Max(float* x, int xSize);
 
-float Min(float* x, size_t xSize);
+float Min(float* x, int xSize);
 
 void NormalizeN(std::vector<float>& input, std::vector<int>& minMax);
 
 void DeNormalizeN(std::vector<float>& input, std::vector<int>& minMax);
 
-void Compress(float* input, size_t inputSize, int* minMax);
+void Compress(float* input, int inputSize, int* minMax);
 
-void DeCompress(std::vector<float>& input, std::vector<size_t>& minMax);
+void DeCompress(std::vector<float>& input, std::vector<int>& minMax);
 
-float VarianceOfPopulation(float* x, size_t size);
+float VarianceOfPopulation(float* x, int size);
 
-float StandardDeviationOfPopulation(float* x, size_t size);
+float StandardDeviationOfPopulation(float* x, int size);
 
-void Standartize(float* dataset, size_t size);
+void Standartize(float* dataset, int size);
 
-void StandartizeLinearContract(float* dataset, size_t datasetSize, int* minMax, float& start, float& end);
+void StandartizeLinearContract(float* dataset, int datasetSize, int* minMax, float& start, float& end);
 
-void DeStandartizeLinearContract(float* dataset, size_t datasetSize, int* minMax);
+void DeStandartizeLinearContract(float* dataset, int datasetSize, int* minMax);
 
-void Normalize(float* dataset, size_t datasetSize);
+void Normalize(float* dataset, int datasetSize);
 
 #endif // STATISTICFUNCTIONS_H
